samples/avm_scheduler/requirements.txt
fastapi
uvicorn
pydantic
httpx
websockets
python-dateutil
openai
python-dotenv
azure-ai-agents
azure-identity